fix: fix redirectflow page wait for provider changed before redirect
